refactor(daily-plugin): split lib.rs; fix sidebar entry awareness and iframe sizing
Split the 1400-line `lib.rs` into focused modules (`state`, `paths`,
`storage`, `markdown_io`, `links`, `indices`, `migration`, `permissions`,
`commands`) so the plugin is easier to extend. No behavior change in the
refactor; existing unit tests move alongside the code they cover.

Fix bug where the daily plugin never learned about the currently-open
entry unless the user clicked in its calendar: `LeftSidebar` was not
forwarding `currentEntry` to `PluginSidebarPanel`, so the iframe's
`init` and `entry-changed` messages always carried `entry: null`. Pass
`entry={currentEntry}` through.

Fix bug where the daily sidebar iframe was wider/taller than its
content, producing spurious scrollbars and a dead gutter. `PluginIframe`
now honors a `{ type: "content-size", height }` message from guests,
dropping its `flex-1` sizing and sizing the iframe to reported content
height so the surrounding sidebar column handles scrolling. The daily
panel opts in by removing `height: 100vh` and the 260px `.panel`
max-width, then posting size via `ResizeObserver`. Other iframe plugins
are unaffected (they fall through to the existing `flex-1` path).
