+++ apps/web/src/lib/components/README.md
@@ -29,7 +29,7 @@ Reusable Svelte components for the editor and UI.
 | `HtmlBlockNodeView.svelte` | Raw HTML block preview/editor shell; preview sanitization now also rewrites `<picture><source srcset>` image candidates in addition to `<img src>` so local images render correctly in Tauri and web preview flows. |
 | `MoreStylesPicker.svelte` | Overflow formatting menu (strikethrough, inline code, spoiler) |
 | `PluginStatusItems.svelte` | Footer status-bar renderer for plugin-contributed items; fetches status once on mount and then refreshes on `SyncStatusChanged`/`SyncProgress` filesystem events instead of polling, eliminating the previous 3-second interval that flooded the IPC channel. |
-| `PluginIframe.svelte` | Sandboxed plugin iframe host with direct `get_component_html` loading when the runtime exposes it, command bridge fallback for older/plugin-command-only guests, managed-context injection for `diaryx.ai`, and response-shape normalization for plugin HTML payloads. |
+| `PluginIframe.svelte` | Sandboxed plugin iframe host with direct `get_component_html` loading when the runtime exposes it, command bridge fallback for older/plugin-command-only guests, managed-context injection for `diaryx.ai`, and response-shape normalization for plugin HTML payloads. Guests may opt into content-sizing by posting `{ type: "content-size", height }` — the host then drops `flex-1` and sizes the iframe to the reported height so the surrounding column (not the iframe) handles scrolling. |
 | `PluginSidebarPanel.svelte` | Sidebar wrapper for plugin component refs (Builtin/Declarative/Iframe). |
 | `SpotlightOverlay.svelte` | Marketplace onboarding spotlight overlay; geometry, tooltip placement, and swipe interpretation live in `spotlightOverlay.ts` so the component shell stays thin and the interaction rules can be unit-tested directly. |
 | `UpgradeBanner.svelte` | Shared upsell surface for Plus-gated features (used for managed AI gating). |
